fix(bindAll): fix constructor not being called with new
BREAKING CHANGE: Properties will use the bind decorator to apply getter/setters
on the prototype instead of the instance. This could cause
issues with consumers currently using the implementain that
assigns the bound properties to the instance through the constructor
rather than on the prototype. The value on the prototype will now be
a getter instead of the original function value.